requirements.txt
absl-py==2.1.0
actionlib==1.14.0
addict==2.4.0
angles==1.9.13
annotated-types==0.7.0
asttokens
attrs==23.2.0
backcall
blinker==1.7.0
bondpy==1.8.6
Brotli
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
certifi
chamferdist==1.0.1
charset-normalizer
click==8.1.7
cloth_simulation_filter==1.1.5
cmake==3.29.2
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
dash==2.16.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy
decorator
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.9.0
docutils==0.21.2
dynamic-reconfigure==1.7.3
einops==0.8.0
entrypoints
exceptiongroup
executing
faiss-gpu==1.7.2
fastjsonschema==2.19.1
filelock
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.3.1
future==1.0.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
geomloss==0.2.6
gitdb==4.0.11
GitPython==3.1.43
gmpy2
gnupg==2.3.1
grpcio==1.62.0
h5py==3.10.0
huggingface-hub==0.24.5
idna
image-geometry==1.16.2
imageio==2.34.2
importlib-metadata==7.0.1
importlib_resources==6.1.2
interactive-markers==1.12.0
ipdb==0.13.13
ipykernel
ipython
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi
Jinja2
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client 
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiss-icp==0.4.0
kiwisolver==1.4.5
laser_geometry==1.6.7
laspy==2.5.4
lazy_loader==0.4
lit==18.1.4
llvmlite==0.43.0
loguru==0.7.2
map_closures 
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe 
matplotlib==3.8.3
matplotlib-inline 
mdurl==0.1.2
message-filters==1.16.0
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
mpi4py
mpmath
msgpack==1.0.8
msgpack-numpy==0.4.8
multimethod==1.11.2
natsort==8.4.0
nbformat==5.9.2
nest_asyncio
networkx
numba==0.60.0
numpy 
numpy-indexed==0.3.7
open3d==0.18.0
open3d-python==0.3.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==23.2
pandas==2.2.1
parso 
pexpect
pickleshare
pillow==10.4.0
platformdirs 
plotly==5.19.0
plyfile==1.0.3
prompt-toolkit 
protobuf==4.25.3
psutil 
ptyprocess 
pure-eval
py-cpuinfo==9.0.0
pycryptodomex==3.20.0
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
Pygments
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks
python-dateutil
python-dotenv==1.0.1
python-qt-binding==0.4.4
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.33.0
regex==2024.7.24
requests==2.32.3
resource_retriever==1.12.7
retrying==1.3.4
rich==13.7.1
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.1
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rpds-py==0.18.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.4.1.post1
scipy==1.13.1
seaborn==0.13.2
sensor-msgs==1.13.1
shapely==2.0.3
shellingham==1.5.4
six
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
smmap==5.0.1
stack-data 
sympy
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorpack==0.11
termcolor==2.4.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.3.0
tifffile==2024.6.18
timm==0.9.16
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
topic-tools==1.16.0
torch==2.0.1
torchaudio==2.0.2
torchpack==0.3.1
torchsparse @ git+https://github.com/mit-han-lab/torchsparse.git@74099d10a51c71c14318bce63d6421f698b24f24
torchvision==0.15.2
tornado
tqdm==4.66.2
traitlets==5.14.1
transformers==4.44.0
triton==2.0.0
typer==0.12.3
typing_extensions
tzdata==2024.1
ultralytics==8.2.51
ultralytics-thop==2.0.0
urllib3
wcwidth
Werkzeug==3.0.1
widgetsnbextension==4.0.10
xacro==1.14.18
yacs==0.1.8
zipp==3.17.0
